Accept secureProtocol as a TLS option for Needle requests
It is possible to run into an issue such as this that will prevent requests against a buggy server from working:

http://carnivore.it/2011/10/07/error_14077458_ssl_routines_ssl23_get_server_hello_reason_1112

The solution is to force SSLv3. This can be done with a mostly undocumented feature of Node.js by doing something like the following:

var options = {
  ...
  secureProtocol: 'SSLv3_method',
  agent: false
}
var req = https.request(options, function(res) {
  ...
});

This change to Needle allows that option to be passed through.