Use debug module
To enable debugging mode, set the environment variable `DEBUG=needle`.

The debug module is the defacto debugging output module, hence a lot of
other modules use it. Because the Needle module previously just checked
for the presense of the DEBUG environment variable it would go into
debug mode when ever the user used the DEBUG environment variable to
debug any other module - this creates a lot of unwanted noise.

For more information of how to debug using the debug module, check:
https://github.com/visionmedia/debug