Use an if statement to handle absence of charset in Content-Type header
I use vscode when debugging my code, and have pause on exceptions enabled.
This means every time I receive from the remote API I'm testing (which does not set a charset in the content type), Execution gets paused.
Using an if statement will ensure we only set the charset when one is present. It also ensure we don't lose performance when the APIs called will never set a charset.